tests/wpt/web-platform-tests/2dcontext/line-styles/2d.line.invalid.strokestyle.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /2dcontext/tools/gentest.py. -->
<title>Canvas test: 2d.line.invalid.strokestyle</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.invalid.strokestyle</h1>
<p class="desc">Verify correct behavior of canvas on an invalid strokeStyle()</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify correct behavior of canvas on an invalid strokeStyle()");
_addTest(function(canvas, ctx) {

ctx.strokeStyle = 'rgb(0, 255, 0)';
ctx.strokeStyle = 'nonsense';
ctx.lineWidth = 200;
ctx.moveTo(0,100);
ctx.lineTo(200,100);
ctx.stroke();
var imageData = ctx.getImageData(0, 0, 200, 200);
var imgdata = imageData.data;
_assert(imgdata[4] == 0, "imgdata[\""+(4)+"\"] == 0");
_assert(imgdata[5] == 255, "imgdata[\""+(5)+"\"] == 255");
_assert(imgdata[6] == 0, "imgdata[\""+(6)+"\"] == 0");


});
</script>
